Fix broken IO logic in CredSSP negotiation.
I'm stunned this code actually worked before. Two variable length reads
that happened to return all the bytes necessary. Even worse, it was
special-cased to occur when the number of bytes to read was specified as
zero.

A new function is defined that knows how to read an ASN.1
tag-length-value triple from a stream rather than doing a reads of
unspecified length.